Rename setup_serena→setup, move read_project_name to config (#44)
