+++ .uncoded/namespace.yaml
@@ -18,6 +18,7 @@ src/:
       main:
     config.py:
       find_pyproject_toml:
+      read_project_name:
       read_source_roots:
       read_instruction_files:
     extract.py:
@@ -56,11 +57,10 @@ src/:
       SERENA_PROJECT_YML:
       SERENA_ALLOWED_TOOLS:
       _STATUS_VERB:
-      read_project_name:
       _sync_mcp_json:
       _sync_serena_project:
       _sync_claude_settings:
-      setup_serena:
+      setup:
     skill.py:
       SKILL_OUTPUTS:
       LEGACY_SKILL_OUTPUTS:
@@ -140,6 +140,11 @@ tests/:
       test_finds_in_cwd:
       test_finds_in_parent:
       test_returns_none_if_not_found:
+    TestReadProjectName:
+      test_reads_name_from_pyproject_toml:
+      test_falls_back_to_cwd_name_when_no_pyproject:
+      test_falls_back_to_cwd_name_when_no_project_section:
+      test_falls_back_to_cwd_name_when_name_missing:
     TestReadSourceRoots:
       test_reads_source_roots:
       test_raises_if_no_pyproject_toml:
@@ -208,12 +213,7 @@ tests/:
   test_serena_setup.py:
     REPO_ROOT:
     EXPECTED_EXCLUDED_TOOLS:
-    TestReadProjectName:
-      test_reads_name_from_pyproject_toml:
-      test_falls_back_to_cwd_name_when_no_pyproject:
-      test_falls_back_to_cwd_name_when_no_project_section:
-      test_falls_back_to_cwd_name_when_name_missing:
-    TestSetupSerena:
+    TestSetup:
       _run:
       test_creates_all_three_files:
       test_mcp_json_is_valid_and_pins_version:
