+++ .uncoded/stubs/src/uncoded/config.pyi
@@ -8,6 +8,10 @@ def find_pyproject_toml() -> Path | None:
     """Search for pyproject.toml starting from cwd, walking up."""
     ...
 
+def read_project_name() -> str:
+    """Read the project name from pyproject.toml, falling back to the cwd name."""
+    ...
+
 def read_source_roots() -> list[Path]:
     """Read source roots from [tool.uncoded] source-roots in pyproject.toml."""
     ...
