+++ tests/test_config.py
@@ -6,6 +6,7 @@ import pytest
 from uncoded.config import (
     find_pyproject_toml,
     read_instruction_files,
+    read_project_name,
     read_source_roots,
 )
 from uncoded.instruction_files import DEFAULT_INSTRUCTION_FILES
@@ -27,6 +28,27 @@ class TestFindPyprojectToml:
     def test_returns_none_if_not_found(self, tmp_path):
         os.chdir(tmp_path)
         assert find_pyproject_toml() is None
+
+
+class TestReadProjectName:
+    def test_reads_name_from_pyproject_toml(self, tmp_path):
+        (tmp_path / "pyproject.toml").write_text('[project]\nname = "my-app"\n')
+        os.chdir(tmp_path)
+        assert read_project_name() == "my-app"
+
+    def test_falls_back_to_cwd_name_when_no_pyproject(self, tmp_path):
+        os.chdir(tmp_path)
+        assert read_project_name() == tmp_path.name
+
+    def test_falls_back_to_cwd_name_when_no_project_section(self, tmp_path):
+        (tmp_path / "pyproject.toml").write_text("[tool.ruff]\n")
+        os.chdir(tmp_path)
+        assert read_project_name() == tmp_path.name
+
+    def test_falls_back_to_cwd_name_when_name_missing(self, tmp_path):
+        (tmp_path / "pyproject.toml").write_text('[project]\nversion = "0.1"\n')
+        os.chdir(tmp_path)
+        assert read_project_name() == tmp_path.name
 
 
 class TestReadSourceRoots:
